backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
diffusers==0.29.1
transformers==4.44.2
accelerate==0.33.0
safetensors==0.4.4
xformers==0.0.27.post2
controlnet_aux==0.0.7
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
tqdm==4.66.5
pydantic
python-multipart
torch
controlnet-aux
requests
python-dotenv
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]

